week4/organizer-api/requirements.txt
django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
pillow==8.4.0
